test(tls): relocate CR-010 guard tests to integration suite (trust-boundary fix)
The guard_before_alloc_tests module was placed inside src/analyzer/tls.rs
as a #[cfg(test)] block. The trust-boundary CI gate (W11-D2) scans src/
with a raw grep and cannot distinguish #[cfg(test)] call-sites from
production call-sites — so the three _for_testing() calls inside the module
were flagged as violations.

Move all three tests (non_handshake_record_client_drains_without_parse,
non_handshake_record_server_drains_without_parse, handshake_record_reaches_parser)
to tests/tls_analyzer_tests.rs, which is the canonical home for integration
tests that use _for_testing seams. The tests themselves are unchanged;
only the file location and import style differ (crate:: -> wirerust::, super::*
expanded to explicit use declarations at the module top, helper functions
renamed to avoid collisions with the existing test_flow_key helper).

Trust-boundary gate: zero violations in src/ after this move.
